get anagram search working
