+++ djAerolith/lib/macondo_interface.py
@@ -76,7 +76,7 @@ def make_rpc_call(procedure_name, arguments):
     headers = {'Content-Type': 'application/json'}
     # This makes it work locally with more "complex" queries. On prod,
     # we limit it.
-    arguments['authToken'] = 'abcdef'
+    # arguments['authToken'] = 'abcdef'
     data = {
         'jsonrpc': "2.0",
         'method': procedure_name,
